Introduce AdmiralUpgradeWithoutMigrationIT
This integration test is similar to AdmiralUpgradeIT. The only
difference is that the there is no migration of data from the old
Admiral instance to the new Admiral instance. Instead of that, the new
Admiral instance is started with the storage from the old instance.

The test verifies that it is possible to start an upgraded Admiral
instance with the storage of the pre-upgeaded instance and with that
maintain the existing entities (containers, networks, apps, etc.)

Change-Id: I64115e99bf88e6cf095f805f04a1d6c318785a4c
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/36705
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Georgi Muleshkov <gmuleshkov@vmware.com>