Removed using "/" for kubernetes log service
For now "/" was used in the self link of the log document. The whole
container self link was used to generate the self link of the log. This
will not work with the latest xenon. Changed it for the current version
in order to avoid upgrade problems when using the latest xenon.
- Changed the generation of the id to be with only the id if the
container not the self link.
- Updated the tests

Change-Id: I2106ca10d084589bdb742f35576328179ee3dbf4
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/37392
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Antonio Filipov <afilipov@vmware.com>