Disallow the same endpoint to be added in multiple projects/groups
Before this commit, it was possible to add an endpoint with the same API
address to multiple projects/business groups. However, now it is
possible to choose to which business groups a given endpoint is
assigned. Therefore, it is possible to get to the state where two
endpoints with the same API address exist in the same project (example:
add endpoint 1 to project 1. Then add endpoint 2 with the same address
to project 2. Then assign project 1 to endpoint 2).

To fix this, with this commit it will be no longer possible to have the
same endpoint added in a different group/project under the same tenant.
It is also not possible to add an endpoint with the same API address
with a different user. However, having the same endpoint in a different
tenant is fine.

Change-Id: I8e442f1ee2bd32fde730b3fd70086f3939d6d6eb
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/38593
Reviewed-by: Antonio Filipov <afilipov@vmware.com>
Bellevue-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>
Closures-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>
CS-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>
Upgrade-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>