[VBV-2109] Ensure utils.js always passes strings to validator
Recently, we have bumped to a newer version of the validator library
(commit 60bb328336afd3195a09147cb46c654aca8586ef). In this new version,
the library asserts that any input it gets is a string value. If it is
not, it throws an exception. This causes VBV-2109 where a template
cannot be modified after being imported via the UI. The problem is
the validation of the cluster size field. The bug was reproducible in
CAFE mode only and goes like this:
- the UI retrieves the container descriptions from the backend
- the CAFE service returns a JSON object with null values explicitly set
  for unset properties (e.g. cluster size)
- the UI detects data change (null is not undefined) and overwrites the
  default value with null
- when validating the form, an empty cluster size is treated as
  numerical 1. This was OK in the previous version of validator but is
  not now.

To address this, the following changes are implemented:
- the utils functions in utils.js now append '' to the argument before
  calling validator functions
- when looking for data changes in the container definition form, null
  and undefined are now treated as the same value
- an additional change in the CAFE services could be made to skip null
  values in responses (not part of this review)

Change-Id: I7078201a3e1d5bdb5e4e86377f6b93a0fe527b78
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/41549
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
PG-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Iveta Ilieva <iilieva@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>