[Bugzilla-2185308] Change immutable map for muttable in ContainerToNetworkAffinityHostFilter
In ContainerToNetworkAffinityHostFilter there was a special case where
we were constructing an immutable map. That map was passed down the
chain of host filters and could cause an "Unexpected error, status: 0"
error message when provisioning an application with a network and an
existing volume. This error was shown because of a unsupported operation
exception thrown by the immutable map when the NamedVolumeAffinityHostFilter
tried to modify it.

The commit replaces the immutable map with a muttable one. It also adds
an integration test to verify that the issue will not happen again.

Also, additional testing to increase code coverage above limit.

Change-Id: I49c2f0c76fb978bbc3e6eb8dfda55668218b1e17
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/42888
Closures-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>