VBV-2212 - Incorrect memory usage information on rescan
The problem was that on rescan the max/total
memory for the placement was updated using the logic : if add operation
add the memory of the host to the total in case of remove: decrease the
total memory. Because of this during rescan we were observing
double, triple and so on amount of memory for the PKS cluster.
- Added a flag in the UI in order to make the difference when it is a
rescan operation.

Change-Id: I9c402a62ec336c68832e25c759b5007d6f74ebff
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/43223
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
PG-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>