[VBV-2209] Display request type for requests in all related reviews
As part of VBV-2209, additional metadata (namely the type of the
request) is going to be stored in each request status. This change
modifies the entries in the request statuses list to display as much
info as is available for a given request.

Change-Id: I7be3679f6483596a9eaf584373b8f2fac5f9e882
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/43625
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Iveta Ilieva <iilieva@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>