[VBV-2220] Fix for PKS clusters with space in the name
When sending requests to PKS API and the cluster name contains spaces
they are not encoded as %20 despite the constructed URI instance is
correct (with %20). This seems to be a xenon bug as it is getting the
path which is then back decoded to ' ', instead of using the rawPath.
Implementing a workaround by encoding the ' ' to '%2520' so xenon would
get '%20'.

Change-Id: Ia54603a7b35829eeb1441930501a0d8cdfbe10de
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/43632
Upgrade-Verified: jenkins <jenkins@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Georgi Muleshkov <gmuleshkov@vmware.com>