[VBV-2218] Workaround for serializing pks cluster
- Created workaround for GSON bug. It fails serialize pks clusters
when one of them has a space in the master host name. This is an
incorrect state but PKS allows it.
- This workaround should be removed when the Xenon gets bumped
to 1.6.14 because it has a newer version of GSON as a dependency.
In that version GSON doesn't have this problem any more.

Change-Id: I91ebad02006216c41e5b6fb1be4508197e0d6492
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/43740
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Alek Boninski <boninskia@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>