[VBV-1886] Denied delete for unauthorized users
- Added check in AbstractTaskStatefulService if the user is
permitted to do the action. This solves the problem that first the
task is started and then is the authorization check.
- Added interceptor that catches 'DELETE' action and fails the
operation if the user is not CLOUD ADMIN. This is done in order to
avoid introducing new role that has access for all other action
but not for 'DELETE'.
- Added unit tests for both changes. For the unit tests:
	- Added dependency in module 'host' of 'auth' in order to
extend the RoleRestrictionsTest and reuse it's funtionallity and have
access to the other classes and the interceptor at the same time.
	- Added the 4 roles that were missing in
host/src/test/resources/local-users.json in order to use them in the
tests and avoid unnecessary duplication.
	- In RequestStatusServiceTest.verifyRequestStatus removed the
code that checks document versions because postgres doesn't keep them
and that assert will always fail when run with postgres.

Change-Id: I971fa7d60a5dd6f6601e7d3f007efe776a62dc2a
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/44424
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
PG-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Stanislav Hadjiiski <hadjiiskis@vmware.com>