VBV-2234 - [Running unit tests locally without retry] ContainerRemovalTaskServiceTest.testRemovingOfCompositeDescriptionAndContainerRemovals
The problem was that we had a couple of operations executing in parallel
when performing remove of a container: remove container state, update
placement, release ports...
During this sequence it was possible to have the case where a delete and
get to the container description are sent at almost the same time. In
this case we hit an issue in xenon and the description is not deleted
even when the delete request is successful. For more information about
the xenon issue check: https://review.ec.eng.vmware.com/#/c/17003/. The
issue is still reproducible with the new xenon
- Changed the removal task first to perform placement and port
operations and after that to delete the container state and the
description
- Checked that the case is fixed by running the test 100 times in a loop
with no failures

Change-Id: I3b420a58077ba7ff5f0aa07fddb043b88f6fe2b4
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/44425
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
PG-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>