Fix Randomly failing testRemoveApplicationWithScaledContainer
The problem was in this test we send delete to the container description
twice. The first request is processed but the second and the second in
in the queue. After the first delete is processed the other gets
cancellation exception. This is the behavior of xenon.
- Added retry in case we get cancellation. We will retry once only in
case of cancellation exception. If the document is deleted and delete is
sent we get 200.
- Fixed a problem in container service. There was a retry but it was not
limited

Change-Id: Ic14711bd5071b8fbb63eb55d1a7d01d2926bb024
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/45353
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
PG-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>