VBV-2239 Adding a k8s cluster succeeds without required credentials
This is because newer k8s apiserver versions (>= 1.8.x) apparently do not
require authentication for api calls on /healthz. When we add a cluster
we do a GET on the /healthz api and we get 200 even though no credentials
are provided but they might be required.

Apparently if the request contains an invalid Bearer token, the response
is actually 401 but if the header is omitted the response is 200 so it
might be a bug in the k8s api server.

The fix is to change /healthz to /healthz/ping since the latter requires
authentication.

Change-Id: Ie2de8a9dd9e4972e0691b9742b22c6833a4b3eaf
Reviewed-on: https://bellevue-ci.eng.vmware.com:8080/45542
Closures-Verified: jenkins <jenkins@vmware.com>
PG-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>