VBV-2270 - [Randomly failing tests] ContainerRemovalTaskService.testRemoveOfStaleContainerOperationFromHostRemoval with postgres
The problem was that the test was sending delete to the container state
and right after that it was asserting that the description is null. Like
this more or less at the same time get and delete were sent to the
container description service and this issue was hit:
https://review.ec.eng.vmware.com/#/c/17003/

The fix:
Container service on delete used to first complete the operation and
after that delete the description.
- Changed the service not to complete before deleting the description in
order to avoid the case described above.

Change-Id: I60e2127636354d82e230f9c9e7c6e4106f885b03