[VBV-1849][VBV-2382]Drop shell in a box
Part of the admiral's agent is a go app that uses the "shell in a box"
utility to expose console access to managed containers. Sadly, this is a
causing some security issues and needs to be removed. On the bright
side(?), it was never officially supported.

This commit drops the whole go app and forces a 'false' value for the
console feature toggle both in UI and backend. The dead code could be
removed in future commits or could be reused when we find a more secure
utility.

The admiral agent is still needed because it is used to query host stats
and to distribute registry/host certificates. However, it no longer
exposes any ports and thus has no SSL certificates.

Change-Id: I0999dba980fad5d9e31f114c282959e78a361586
Reviewed-on: https://bellevue-ci-gerrit.eng.vmware.com:443/c/64050
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
PG-Verified: jenkins <jenkins@vmware.com>