[Bugzilla-2325149]Publish All Ports option is always forced to true
Before this commit, the initial state of the checkbox was "selected".
Then, when the data for the container description gets
loaded, that data is compared to the local state. The loaded data
contains publishAll = false and the local state contains no publishAll
property at all. Since our update check treats undefined and false as
the same value, we decide that no change was made and keep the checkbox
selected.

To resolve this and also to keep the current defaults (publish all is on
by default), this commit sets the default value for the publishAll
property in the description's state. These cases are covered:
- Adding a new container to a template
- Adding additional info to an image in repositories view (creating a
  template from image)
- Directly provisioning a container from an image

In all of the cases mentioned above, publish all was and still is on by
default.

Change-Id: I1b0f2978ae7c79ffb28bbbf17269bc2ef4c3153c
Reviewed-on: https://bellevue-ci-gerrit.eng.vmware.com:443/c/65588
Reviewed-by: Iveta Ilieva <iilieva@vmware.com>
Closures-Verified: jenkins <jenkins@vmware.com>
Upgrade-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
PG-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>