[Bug 2484719] Stop host in case of auth provider init error
Changed behaviour to stop the host in case of initialization error of
the authentication provider. If any of initBootConfig or initConfig fail
the error is logged and host is stopped.
Keeping the previous behaviour, the operation to the handleStart is
completed immediately, started separate thread waiting to complete the
initialization and stop the host if necessary.

Change-Id: Idca3890dc6ec043a439d538414af83902968aff9
Reviewed-on: https://bellevue-ci-gerrit.eng.vmware.com:443/c/112141
Upgrade-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
PG-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>